Googletest export
Extend gtest-port and stubs for ESP_PLATFORM

ESP_PLATFORM is the macro used to indicate compilation for the ESP32
using the esp-idf. This isn't a fully posix compatible system so
various features of google test need to be stubbed out in order for
it to work. It's oddly similar to the GTEST_OS_WINDOWS_PHONE setup.

PiperOrigin-RevId: 267471968