Use GTEST_DONT_DEFINE_TEST_F to guard TEST_F
The documentation is clear that the FOO we'll be guarding always matches
the spelling of the DONT macro.  A single guard macro should not toggle
more than one implementation macro.

This fixes a regression in 7413280c52c1f759395572a384165023d24eeb57.

Relatedly, improve the documentation of the DONT macros to bring the
list of valid FOO values up to date.